Arbolado/requirements/local.txt
Django==3.1.3
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-model-utils==4.0.0
Pillow==7.2.0
psycopg2==2.8.6
Unipath==1.1